+++ waffle-provider/package.json
@@ -12,8 +12,8 @@
   "scripts": {
     "prepublishOnly": "yarn build",
     "test": "export NODE_ENV=test && mocha",
-    "lint": "eslint '{lib,test}/**/*.ts'",
-    "lint:fix": "eslint --fix '{lib,test}/**/*.ts'",
+    "lint": "eslint '{src,test}/**/*.ts'",
+    "lint:fix": "eslint --fix '{src,test}/**/*.ts'",
     "build": "rimraf ./dist && yarn build:esm && yarn build:cjs",
     "build:esm": "tsc -p tsconfig.build.json --outDir dist/esm --module ES6",
     "build:cjs": "tsc -p tsconfig.build.json --outDir dist/cjs --declaration false"
