+++ waffle-provider/src/deprecate.ts
@@ -1,5 +1,5 @@
-export function deprecate (what: string, message?: string) {
+export function deprecate(what: string, message?: string) {
   if (process.env.NODE_ENV !== 'production') {
-    console.warn(`Deprecation warning: ${what} is deprecated.${message ? ' ' + message : ''}`)
+    console.warn(`Deprecation warning: ${what} is deprecated.${message ? ' ' + message : ''}`);
   }
 }
