+++ waffle-waffle/package.json
@@ -17,8 +17,8 @@
     "test:buildonly": "ts-node -T buildTestContracts",
     "test:nobuild": "export NODE_ENV=test && mocha",
     "test": "yarn test:buildonly && yarn test:nobuild",
-    "lint": "eslint '{lib,test}/**/*.ts'",
-    "lint:fix": "eslint --fix '{lib,test}/**/*.ts'",
+    "lint": "eslint '{src,test}/**/*.ts'",
+    "lint:fix": "eslint --fix '{src,test}/**/*.ts'",
     "build": "rimraf ./dist && yarn build:esm && yarn build:cjs",
     "build:esm": "tsc -p tsconfig.build.json --outDir dist/esm --module ES6",
     "build:cjs": "tsc -p tsconfig.build.json --outDir dist/cjs --declaration false",
