+++ waffle-waffle/src/index.ts
@@ -1,7 +1,7 @@
 import './matchers/matchertypes';
 
-export * from '@waffle/provider'
-export * from './deployContract'
+export * from '@waffle/provider';
+export * from './deployContract';
 export * from './link';
 export * from './fixtures';
 export {default as solidity} from './matchers/matchers';
